+++ packages/core/forms/src/components/forms/OIDCPrincipals.vue
@@ -28,7 +28,7 @@
         @change="handleModeChange"
       >
         <div class="auth-mode-card-content">
-          <KeyIcon :size="KUI_ICON_SIZE_40" />
+          <KeyIcon :size="`var(--kui-icon-size-40, ${KUI_ICON_SIZE_40})`" />
           <div class="auth-mode-card-label">
             Kong Identity
           </div>
@@ -46,7 +46,7 @@
         @change="handleModeChange"
       >
         <div class="auth-mode-card-content">
-          <WorldIcon :size="KUI_ICON_SIZE_40" />
+          <WorldIcon :size="`var(--kui-icon-size-40, ${KUI_ICON_SIZE_40})`" />
           <div class="auth-mode-card-label">
             External auth server
           </div>
@@ -76,7 +76,7 @@
             data-testid="create-auth-server-action"
             @click.stop="leavePromptType = 'authServer'"
           >
-            <AddIcon :size="KUI_ICON_SIZE_20" />
+            <AddIcon :size="`var(--kui-icon-size-20, ${KUI_ICON_SIZE_20})`" />
             <div class="create-action-content">
               <span>Create authorization server</span>
               <div class="create-action-hint">
@@ -111,7 +111,7 @@
               data-testid="create-client-action"
               @click.stop="leavePromptType = 'client'"
             >
-              <AddIcon :size="KUI_ICON_SIZE_20" />
+              <AddIcon :size="`var(--kui-icon-size-20, ${KUI_ICON_SIZE_20})`" />
               <div class="create-action-content">
                 <span>Create client</span>
                 <div class="create-action-hint">
@@ -149,7 +149,7 @@
           type="button"
           @click="removeClientRow(index)"
         >
-          <CloseIcon :size="KUI_ICON_SIZE_30" />
+          <CloseIcon :size="`var(--kui-icon-size-30, ${KUI_ICON_SIZE_30})`" />
         </button>
       </div>
 
@@ -320,7 +320,7 @@
           type="button"
           @click="removeClientRow(index)"
         >
-          <CloseIcon :size="KUI_ICON_SIZE_30" />
+          <CloseIcon :size="`var(--kui-icon-size-30, ${KUI_ICON_SIZE_30})`" />
         </button>
       </div>