+++ packages/entities/entities-plugins/src/components/free-form/plugins/datakit/flow-editor/node/FlowNode.vue
@@ -26,7 +26,7 @@
       <WarningIcon
         v-if="error"
         class="error-icon"
-        :color="KUI_COLOR_TEXT_DANGER"
+        :color="`var(--kui-color-text-danger, ${KUI_COLOR_TEXT_DANGER})`"
         :size="16"
       />
       <KDropdown
@@ -47,7 +47,7 @@
           icon
           size="small"
         >
-          <MoreIcon :color="KUI_COLOR_TEXT" />
+          <MoreIcon :color="`var(--kui-color-text, ${KUI_COLOR_TEXT})`" />
         </KButton>
         <template #items>
           <KDropdownItem
@@ -115,12 +115,12 @@
               <template v-if="hasInputFields">
                 <UnfoldMoreIcon
                   v-if="!inputsExpanded"
-                  :size="KUI_ICON_SIZE_20"
+                  :size="`var(--kui-icon-size-20, ${KUI_ICON_SIZE_20})`"
                 />
                 <UnfoldLessIcon
                   v-if="inputsExpanded"
-                  :color="inputsCollapsible ? undefined : KUI_COLOR_TEXT_DISABLED"
-                  :size="KUI_ICON_SIZE_20"
+                  :color="inputsCollapsible ? undefined : `var(--kui-color-text-disabled, ${KUI_COLOR_TEXT_DISABLED})`"
+                  :size="`var(--kui-icon-size-20, ${KUI_ICON_SIZE_20})`"
                 />
               </template>
             </div>
@@ -187,12 +187,12 @@
               <template v-if="hasOutputFields">
                 <UnfoldMoreIcon
                   v-if="!outputsExpanded"
-                  :size="KUI_ICON_SIZE_20"
+                  :size="`var(--kui-icon-size-20, ${KUI_ICON_SIZE_20})`"
                 />
                 <UnfoldLessIcon
                   v-if="outputsExpanded"
-                  :color="outputsCollapsible ? undefined : KUI_COLOR_TEXT_DISABLED"
-                  :size="KUI_ICON_SIZE_20"
+                  :color="outputsCollapsible ? undefined : `var(--kui-color-text-disabled, ${KUI_COLOR_TEXT_DISABLED})`"
+                  :size="`var(--kui-icon-size-20, ${KUI_ICON_SIZE_20})`"
                 />
               </template>
             </div>
@@ -626,7 +626,7 @@ $one-over-sqrt2-px: math.pow(2, -0.5) * 1px;
       gap: var(--kui-space-30, $kui-space-30);
       justify-self: start;
       max-width: 100%;
-      /* stylelint-disable-next-line @kong/design-tokens/use-proper-token */
+      /* stylelint-disable-next-line @kong/stylelint-plugin-design-tokens/use-proper-token */
       max-width: calc(100% - var(--kui-space-30, $kui-space-30));
       position: relative;
 
@@ -698,7 +698,7 @@ $one-over-sqrt2-px: math.pow(2, -0.5) * 1px;
       &.connecting::after,
       &:hover::after {
         background-color: var(--kui-color-background-primary, $kui-color-background-primary);
-        /* stylelint-disable-next-line @kong/design-tokens/use-proper-token */
+        /* stylelint-disable-next-line @kong/stylelint-plugin-design-tokens/use-proper-token */
         box-shadow: 0 0 0 1px var(--kui-color-background-primary, $kui-color-background-primary);
       }
     }