+++ packages/entities/entities-plugins/src/components/select/PluginCatalogListView.vue
@@ -99,8 +99,8 @@
               </KDropdownItem>
             </template>
             <MoreIcon
-              :color="KUI_COLOR_TEXT_NEUTRAL"
-              :size="KUI_ICON_SIZE_30"
+              :color="`var(--kui-color-text-neutral, ${KUI_COLOR_TEXT_NEUTRAL})`"
+              :size="`var(--kui-icon-size-30, ${KUI_ICON_SIZE_30})`"
             />
           </KDropdown>
         </div>